+++ src/main/java/org/swyg/greensumer/repository/ReviewPostEntityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.swyg.greensumer.domain.ReviewPostEntity;
 
-public interface ReviewPostEntityRepository extends JpaRepository<ReviewPostEntity, Integer> {
+public interface ReviewPostEntityRepository extends JpaRepository<ReviewPostEntity, Long> {
 
-    Page<ReviewPostEntity> findAllByUser_Id(Integer userId, Pageable pageable);
+    Page<ReviewPostEntity> findAllByUser_Id(Long userId, Pageable pageable);
 }
